Fixes the broken Windows builds by adding --use-loca-env to nvcc calls.
conda-forge/nvcc-feedstock#76 This library uses an old version of CMake
CUDA find_library and a strange way of declaring the build artifacts, so
the flag needs to be added via -DCUDA_NVCC_FLAGS="--use-local-env" instead
of the CUDA_FLAGS environment variable or by appending the option to the
CUDACXX environment variable.

Changes the build and install steps to call cmake instead of
generator/build system directly. Switches generator to Ninja for all
platforms. Based on this PR, Ninja builds seem significantly faster or
approximately the same duration.

Adds a special case for 11.2 in the Windows build script because it was missing.

Adds checks for additional build artifacts.

Adds a strict run_export.

Changes the maintainers from @isuruf to myself and @conda-forge/pytorch-cpu.